delete unnecessary blockchain-scanner.service*
