--- libs/account-lib/src/utils/blockchain-scanner.service.spec.ts
@@ -1,205 +1,0 @@
-import { Injectable, Logger } from '@nestjs/common';
-import { Test, TestingModule } from '@nestjs/testing';
-import { BlockHash, BlockNumber, SignedBlock } from '@polkadot/types/interfaces';
-import { DEFAULT_REDIS_NAMESPACE, getRedisToken, InjectRedis } from '@songkeys/nestjs-redis';
-import { Redis } from 'ioredis';
-import { BlockchainScannerService } from './blockchain-scanner.service';
-import { FrameSystemEventRecord } from '@polkadot/types/lookup';
-import { mockApiPromise } from '#testlib/polkadot-api.mock.spec';
-import { BlockchainService, NONCE_SERVICE_REDIS_NAMESPACE } from '#blockchain/blockchain.service';
-import { IBlockchainNonProviderConfig } from '#blockchain/blockchain.config';
-import { GenerateMockConfigProvider } from '#testlib/utils.config-tests';
-import { EventEmitterModule } from '@nestjs/event-emitter';
-import { AnyNumber } from '@polkadot/types/types';
-import { InjectPinoLogger, LoggerModule } from 'nestjs-pino';
-import { getPinoHttpOptions } from '#logger-lib';
-
-jest.mock('@polkadot/api', () => {
-  const originalModule = jest.requireActual<typeof import('@polkadot/api')>('@polkadot/api');
-  return {
-    __esModules: true,
-    WsProvider: jest.fn().mockImplementation(() => ({
-      disconnect: jest.fn(),
-      ...originalModule.WsProvider,
-    })),
-    ApiPromise: jest.fn().mockImplementation(() => ({
-      consts: {},
-      disconnect: jest.fn(),
-      ...originalModule.ApiPromise,
-      ...mockApiPromise,
-    })),
-  };
-});
-const mockBlockchainConfigProvider = GenerateMockConfigProvider<IBlockchainNonProviderConfig>('blockchain', {
-  frequencyTimeoutSecs: 10,
-  frequencyApiWsUrl: new URL('ws://localhost:9944'),
-  isDeployedReadOnly: false,
-});
-
-const mockRedis = {
-  get: jest.fn(),
-  set: jest.fn(),
-  defineCommand: jest.fn(),
-};
-
-const mockDefaultRedisProvider = {
-  provide: getRedisToken(DEFAULT_REDIS_NAMESPACE),
-  useValue: mockRedis,
-};
-
-const mockNonceRedisProvider = {
-  provide: getRedisToken(NONCE_SERVICE_REDIS_NAMESPACE),
-  useValue: mockRedis,
-};
-
-const mockBlockHash = {
-  toString: jest.fn(() => '0x1234'),
-  some: () => true,
-  isEmpty: false,
-} as unknown as BlockHash;
-
-const mockSignedBlock = {
-  block: {
-    header: {
-      number: 1,
-    },
-    extrinsics: [],
-  },
-};
-
-const mockEmptyBlockHash = {
-  toString: jest.fn(() => '0x00000'),
-  some: () => false,
-  isEmpty: true,
-} as unknown as BlockHash;
-
-@Injectable()
-class ScannerService extends BlockchainScannerService {
-  constructor(
-    @InjectRedis() redis: Redis,
-    blockchainService: BlockchainService,
-    @InjectPinoLogger('ScannerService') protected readonly logger,
-  ) {
-    super(redis, blockchainService, logger);
-  }
-
-  // eslint-disable-next-line
-  protected processCurrentBlock = jest.fn((_currentBlock: SignedBlock, _blockEvents: FrameSystemEventRecord[]) => {
-    return Promise.resolve();
-  });
-}
-
-describe('BlockchainScannerService', () => {
-  let service: ScannerService;
-  let blockchainService: BlockchainService;
-  let moduleRef: TestingModule;
-
-  beforeAll(async () => {
-    moduleRef = await Test.createTestingModule({
-      imports: [
-        LoggerModule.forRoot(getPinoHttpOptions()),
-        EventEmitterModule.forRoot({
-          // Use this instance throughout the application
-          global: true,
-          // set this to `true` to use wildcards
-          wildcard: false,
-          // the delimiter used to segment namespaces
-          delimiter: '.',
-          // set this to `true` if you want to emit the newListener event
-          newListener: false,
-          // set this to `true` if you want to emit the removeListener event
-          removeListener: false,
-          // the maximum amount of listeners that can be assigned to an event
-          maxListeners: 10,
-          // show event name in memory leak message when more than maximum amount of listeners is assigned
-          verboseMemoryLeak: false,
-          // disable throwing uncaughtException if an error event is emitted and it has no listeners
-          ignoreErrors: false,
-        }),
-      ],
-      providers: [
-        mockDefaultRedisProvider,
-        mockNonceRedisProvider,
-        mockBlockchainConfigProvider,
-        Logger,
-        BlockchainService,
-        ScannerService,
-      ],
-    }).compile();
-    moduleRef.enableShutdownHooks();
-    service = moduleRef.get<ScannerService>(ScannerService);
-    blockchainService = moduleRef.get<BlockchainService>(BlockchainService);
-    const mockApi: any = await blockchainService.getApi();
-    jest.spyOn(blockchainService, 'getBlock').mockResolvedValue(mockSignedBlock as unknown as SignedBlock);
-    jest
-      .spyOn(blockchainService, 'getBlockHash')
-      .mockImplementation((blockNumber: BlockNumber | AnyNumber) =>
-        Promise.resolve((blockNumber as unknown as number) > 1 ? mockEmptyBlockHash : mockBlockHash),
-      );
-    jest.spyOn(blockchainService, 'getLatestBlockNumber');
-    jest.spyOn(blockchainService, 'getEvents').mockResolvedValue([]);
-    mockApi.emit('connected'); // keeps the test suite from hanging when finished
-  });
-
-  afterAll(async () => {
-    moduleRef.close();
-  });
-
-  describe('scan', () => {
-    it('Should call processCurrentBlock', async () => {
-      const processBlockSpy = jest.spyOn(service as unknown as any, 'processCurrentBlock');
-      await service.scan();
-      expect(processBlockSpy).toHaveBeenCalled();
-    });
-  });
-
-  describe('checkScanParameters', () => {
-    const latestFinalizedBlockNumber = 8;
-
-    beforeEach(() => {
-      jest.restoreAllMocks();
-      service.scanParameters = { onlyFinalized: true };
-      jest.spyOn(blockchainService, 'getLatestBlockNumber').mockResolvedValue(latestFinalizedBlockNumber);
-    });
-
-    afterAll(() => {
-      jest.restoreAllMocks();
-    });
-
-    describe('proccessing only finalized blocks', () => {
-      beforeEach(() => {
-        service.scanParameters = { onlyFinalized: true };
-      });
-
-      it('encounter a non-finalized block should throw', async () => {
-        await expect(
-          (service as unknown as any).checkScanParameters(latestFinalizedBlockNumber + 1, mockBlockHash),
-        ).rejects.toThrow(/^Latest finalized block/);
-      });
-
-      it('encounter a finalized block should not throw', async () => {
-        await expect(
-          (service as unknown as any).checkScanParameters(latestFinalizedBlockNumber, mockBlockHash),
-        ).resolves.toBeFalsy();
-      });
-    });
-
-    describe('allowing unfinalized blocks', () => {
-      beforeEach(() => {
-        service.scanParameters = { onlyFinalized: false };
-      });
-
-      it('encounter an empty block should throw', async () => {
-        await expect(
-          (service as unknown as any).checkScanParameters(latestFinalizedBlockNumber, mockEmptyBlockHash),
-        ).rejects.toThrow(/^Empty block/);
-      });
-
-      it('encounter a non-finalized block should not throw', async () => {
-        await expect(
-          (service as unknown as any).checkScanParameters(latestFinalizedBlockNumber + 1, mockBlockHash),
-        ).resolves.toBeFalsy();
-      });
-    });
-  });
-});
